requirements.txt
langchain==0.2.13
langchain-core==0.2.34
langchain-openai==0.1.7
streamlit==1.39.0
python-dotenv==1.0.1
requests==2.32.5
